+++ vehicle/myskoda/types.go
@@ -1,4 +1,4 @@
-package publicapi
+package myskoda
 
 import "time"
